refactor class name
