+++ app/src/main/java/com/example/uhyo/online/OnlineJoinRoomFragment.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.uhyo.R
 
-class JoinRoomFragment: Fragment() {
+class OnlineJoinRoomFragment: Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
