+++ app/src/main/java/com/example/uhyo/online/OnlineMakeRoomFragment.kt
@@ -1,4 +1,4 @@
-package com.example.uhyo.main
+package com.example.uhyo.online
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.uhyo.R
 
-class SoloFragment : Fragment(){
+class OnlineMakeRoomFragment: Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -19,6 +19,6 @@ class SoloFragment : Fragment(){
         savedInstanceState: Bundle?
     ): View? {
         super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_main_solo, container, false)
+        return inflater.inflate(R.layout.fragment_online_makeroom, container, false)
     }
 }
