+++ app/src/main/java/com/example/uhyo/online/OnlineSetNameFrament.kt
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.uhyo.R
 
-class QuickMatchFragment: Fragment() {
+class OnlineSetNameFrament: Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -19,6 +19,6 @@ class QuickMatchFragment: Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         super.onCreateView(inflater, container, savedInstanceState)
-        return inflater.inflate(R.layout.fragment_online_quickmatch, container, false)
+        return inflater.inflate(R.layout.fragment_online_setname, container, false)
     }
 }
